Removed int16_t case

+++ src/tap/control/control_operator_interface.hpp
@@ -41,8 +41,8 @@ public:
     static constexpr int16_t USER_MOUSE_PITCH_MAX = 1000;
     static constexpr float USER_MOUSE_YAW_SCALAR = (1.0f / USER_MOUSE_YAW_MAX);
     static constexpr float USER_MOUSE_PITCH_SCALAR = (1.0f / USER_MOUSE_PITCH_MAX);
-    static constexpr float CTRL_SCALAR = (1.0f / (int16_t)4);
-    static constexpr float SHIFT_SCALAR = (1.0f / (int16_t)2);
+    static constexpr float CTRL_SCALAR = (1.0f / 4);
+    static constexpr float SHIFT_SCALAR = (1.0f / 2);
     static constexpr float USER_STICK_SENTINEL_DRIVE_SCALAR = 5000.0f;
     static constexpr float CHASSIS_X_KEY_INPUT_FILTER_ALPHA = 0.05f;
     static constexpr float CHASSIS_Y_KEY_INPUT_FILTER_ALPHA = 0.05f;
